Fix program memory reading output type u16->u8

+++ src/mapper/mapper0.rs
@@ -40,7 +40,7 @@ impl Mapper0 {
 }
 
 impl Mapper for Mapper0 {
-  fn mapped_read_cpu_u8(&self, address: u16) -> u16 {
+  fn mapped_read_cpu_u8(&self, address: u16) -> u8 {
     match address {
       0x6000..=0x7FFF => self.get_rom().prg_ram.read(Page::First(EightKb), address - 0x6000),
       0x8000..=0xBFFF => self.get_rom().prg_rom.read(Page::First(SixteenKb), address - 0x8000),
